// apps/io-func-sign-support/src/infra/http/models/Document.ts
/**
 * Do not edit this file it is auto-generated by io-utils / gen-api-models.
 * See https://github.com/pagopa/io-utils
 */
/* eslint-disable  */

import { Id } from "./Id";
import { DocumentMetadata } from "./DocumentMetadata";
import { Timestamp } from "./Timestamp";
import * as t from "io-ts";
import { enumType } from "@pagopa/ts-commons/lib/types";

export enum StatusEnum {
  "WAIT_FOR_UPLOAD" = "WAIT_FOR_UPLOAD",

  "WAIT_FOR_VALIDATION" = "WAIT_FOR_VALIDATION",

  "READY" = "READY",

  "REJECTED" = "REJECTED"
}

// required attributes
const DocumentR = t.interface({
  id: Id,

  metadata: DocumentMetadata,

  status: enumType<StatusEnum>(StatusEnum, "status"),

  created_at: Timestamp
});

// optional attributes
const DocumentO = t.partial({
  url: t.string,

  uploaded_at: Timestamp,

  updated_at: Timestamp,

  rejected_at: Timestamp,

  reject_reason: t.string
});

export const Document = t.exact(
  t.intersection([DocumentR, DocumentO], "Document")
);

export type Document = t.TypeOf<typeof Document>;